:construction: it compile

+++ libraries/lib-advanced-visualization/include/mensia/advanced-visualization/mIRenderer.hpp
@@ -59,7 +59,7 @@ public:
 	void setHistoryDrawIndex(const size_t index);
 	void feed(const float* data);
 	void feed(const float* data, const size_t nSample);
-	void feed(const uint64_t stimDate, const uint64_t stimID) { m_stimulationHistory.emplace_back((stimDate >> 16) / 65536., stimID); }
+	void feed(const uint64_t stimDate, const size_t stimID) { m_stimulationHistory.emplace_back((stimDate >> 16) / 65536., stimID); }
 	void prefeed(const size_t nPreFeedSample);
 
 	float getSuggestedScale();
